#41 Organize IdentityIndex-related code into a subfolder
