+++ src/langium-model-server/identity/identity-index/derivative-indexed-identities.ts
@@ -1,9 +1,9 @@
 import type { AstNode } from 'langium'
-import type * as sem from '../semantic/model'
-import { ValueBasedMap } from '../utils/collections'
+import type * as sem from '../../semantic/model'
+import { ValueBasedMap } from '../../utils/collections'
 import { AbstractIndexedIdentities } from './abstract-indexed-identities'
-import type { DerivativeIdentityName, Identity } from './model'
-import type { ValueBasedOperations } from '../utils/types'
+import type { DerivativeIdentityName, Identity } from '../model'
+import type { ValueBasedOperations } from '../../utils/types'
 
 export class DerivativeIndexedIdentities<T extends AstNode | sem.ArtificialAstNode, NAME extends DerivativeIdentityName, ID extends Identity<T, NAME>> extends AbstractIndexedIdentities<T, NAME, ID> {
