+++ src/langium-model-server/identity/identity-index/index.ts
@@ -1,0 +1,4 @@
+export * from './abstract-indexed-identities'
+export * from './ast-indexed-identities'
+export * from './derivative-indexed-identities'
+export * from './identity-index'
