+++ src/langium-model-server/langium-model-server-module.ts
@@ -3,7 +3,7 @@ import { LmsDocumentHighlightProvider } from './lsp/document-highlight-provider'
 import { LmsLanguageServer } from './lsp/lms-language-server'
 import { LmsRenameProvider } from './lsp/rename-provider'
 import type { SemanticIdentifier } from './identity/model'
-import type { IdentityIndex } from './identity'
+import type { IdentityIndex } from './identity/identity-index'
 import type { LangiumModelServerDefaultServices, LangiumModelServerDefaultSharedServices, LangiumModelServerServices } from './services'
 import { DefaultLangiumSourceModelServer } from './lms/langium-model-server'
 import { DefaultLmsSubscriptions } from './lms/subscriptions'
