+++ src/langium-model-server/lms/facade.ts
@@ -2,7 +2,7 @@ import { type LangiumDocuments, type LanguageMetaData, type MaybePromise } from 
 import type { Connection } from 'vscode-languageserver'
 import { ShowDocumentRequest } from 'vscode-languageserver'
 import { URI } from 'vscode-uri'
-import type { IdentityIndex } from '../identity'
+import type { IdentityIndex } from '../identity/identity-index'
 import type { IdentityManager } from '../identity/manager'
 import type { SemanticIdentifier } from '../identity/model'
 import type { LangiumModelServerServices } from '../services'
