+++ src/langium-model-server/lms/langium-model-server.ts
@@ -2,7 +2,7 @@ import * as fs from 'fs-extra'
 import * as http2 from 'http2'
 import path from 'path'
 import type { SemanticIdentifier } from '../identity/model'
-import type { IdentityIndex } from '../identity'
+import type { IdentityIndex } from '../identity/identity-index'
 import type { LangiumModelServerServices } from '../services'
 import type { LmsDocument } from '../workspace/documents'
 import { LangiumModelServerRouter } from './router'
