+++ src/langium-model-server/lsp/lms-language-server.ts
@@ -3,7 +3,7 @@ import { DefaultLanguageServer, startLanguageServer } from 'langium'
 import type { Connection, InitializeParams, InitializeResult } from 'vscode-languageserver'
 import { FileChangeType } from 'vscode-languageserver'
 import { URI } from 'vscode-uri'
-import type { IdentityIndex } from '../identity'
+import type { IdentityIndex } from '../identity/identity-index'
 import type { SemanticIdentifier } from '../identity/model'
 import { Save } from '../lms/model'
 import type { LangiumModelServerAddedServices } from '../services'
