+++ src/langium-model-server/services.ts
@@ -1,5 +1,5 @@
 import type { DeepPartial, DocumentHighlightProvider, DocumentValidator, LangiumServices, LanguageServer, PartialLangiumServices, RenameProvider } from 'langium'
-import type { IdentityIndex } from './identity'
+import type { IdentityIndex } from './identity/identity-index'
 import type { IdentityManager } from './identity/manager'
 import type { SemanticIdentifier } from './identity/model'
 import type { IdentityStorage } from './identity/storage'
