+++ src/language-server/task-list/lms/task-list-facade.ts
@@ -12,7 +12,7 @@ import * as sem from '../../../langium-model-server/semantic/model'
 import type { LangiumModelServerServices } from '../../../langium-model-server/services'
 import { LmsDocument } from '../../../langium-model-server/workspace/documents'
 import * as ast from '../../generated/ast'
-import type { TaskListIdentityIndex } from '../identity'
+import type { TaskListIdentityIndex } from '../identity/identity-index'
 import * as identity from '../identity/model'
 import * as semantic from '../semantic/model'
 import type { TaskListSemanticDomain } from '../semantic/task-list-semantic-domain'
